move generateEvents
