+++ tools/queuebench/producer.go
@@ -19,6 +19,7 @@ package main
 
 import (
 	"context"
+	"crypto/rand"
 	"fmt"
 	"log"
 
@@ -42,4 +43,15 @@ func produce(ctx context.Context, producer *kafka.Producer, cfg *config) {
 	for {
 		producer.Produce(ctx, record)
 	}
+}
+
+func generateEvent(size int) ([]byte, error) {
+	buf := make([]byte, size)
+
+	_, err := rand.Read(buf)
+	if err != nil {
+		return []byte{}, fmt.Errorf("error while generating random string: %w", err)
+	}
+
+	return buf, nil
 }
